display pseudotime
